requirements.txt
# Core dependencies
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0

# PDF processing
PyMuPDF>=1.21.0
pdf2image>=1.16.0

# OCR engines
easyocr>=1.6.0
texify>=0.2.0

# Table extraction
img2table>=1.0.0

# Document export
python-docx>=0.8.11
pyhwpx>=0.5.0

# LaTeX to MathML (for DOCX equation support)
latex2mathml>=3.0.0

# GUI
PyQt6>=6.5.0

# Machine learning (for handwriting classification)
torch>=1.9.0
torchvision>=0.10.0

# Utilities
tqdm>=4.62.0
pydantic>=2.0.0
